Mirja's Follow-up
"Thx. However, I think it would still be good to clarify which packets can carry Convert message. I assume you are saying now that there are no restrictions and a Convert exchange can go on for multiple RTT before the actually data transmission is started? Is that correct? 

However, I guess in mist cases (and all changes specified in this doc) you only have convert messages in the syn and syn/ack, and potentially also in the 3rd ack of the handshake but that’s only an error and the connection is closed/reset afterwards. If that is correct, I think that would actually be good to say somewhere early in the document."